pyproject.toml
[project]
name = "hr-system"
version = "0.1.0"
description = ""
authors = [
    {name = "Dmitriy",email = "cubarrv@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (==4.2)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "django-filter (<25.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
